+++ tool/build_ios.sh
@@ -1,5 +1,13 @@
-rustup target add aarch64-apple-ios x86_64-apple-ios
+rustup target add aarch64-apple-ios aarch64-apple-ios-sim x86_64-apple-ios
 cargo build --target aarch64-apple-ios --release
+cargo build --target aarch64-apple-ios-sim --release
 cargo build --target x86_64-apple-ios --release
 
-lipo "target/aarch64-apple-ios/release/libisar.a" "target/x86_64-apple-ios/release/libisar.a" -output "libisar_ios.a" -create
+lipo "target/aarch64-apple-ios-sim/release/libisar.a" "target/x86_64-apple-ios/release/libisar.a" -output "target/aarch64-apple-ios-sim/libisar.a" -create
+xcodebuild \
+    -create-xcframework \
+    -library target/aarch64-apple-ios/release/libisar.a \
+    -library target/aarch64-apple-ios-sim/libisar.a \
+    -output isar.xcframework 
+
+zip -r isar_ios.xcframework.zip isar.xcframework