+++ tool/build_macos.sh
@@ -1,3 +1,5 @@
+export MACOSX_DEPLOYMENT_TARGET=10.11
+
 rustup target add aarch64-apple-darwin x86_64-apple-darwin
 cargo build --target aarch64-apple-darwin --release
 cargo build --target x86_64-apple-darwin --release