+++ packages/isar_inspector/lib/error_screen.dart
@@ -1,13 +1,10 @@
+// ignore: avoid_web_libraries_in_flutter
+import 'dart:html';
+
 import 'package:flutter/material.dart';
 
 class ErrorScreen extends StatelessWidget {
-  const ErrorScreen({
-    super.key,
-    required this.message,
-    required this.retry,
-  });
-  final String message;
-  final VoidCallback retry;
+  const ErrorScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -15,18 +12,15 @@ class ErrorScreen extends StatelessWidget {
       child: Column(
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
-          Text(
-            message,
-            style: const TextStyle(fontSize: 20),
+          const Text(
+            'Disconnected',
+            style: TextStyle(fontSize: 20),
           ),
           const SizedBox(height: 10),
-          const Text(
-            'Please make sure your Isar instance is '
-            'running and you use the Chrome browser.',
-          ),
+          const Text('Please make sure your Isar instance is running.'),
           const SizedBox(height: 40),
           ElevatedButton(
-            onPressed: retry,
+            onPressed: window.location.reload,
             child: const Text('Retry Connection'),
           ),
         ],